financial_agent/requirements.txt
qdrant-client
openai
python-dotenv
redis
sentence-transformers
rank_bm25
numpy
requests
beautifulsoup4
pydantic
sec-edgar-api
tenacity
termcolor
flashrank
pymupdf
pillow
langchain-text-splitters
tqdm
